fix: only enqueue editor scripts on editor screen
Fix PHPCS errors
Sync all post meta
