+++ config/webpack.common.js
@@ -1,5 +1,3 @@
-/* global process, module, require */
-
 const path = require( 'path' );
 const CleanWebpackPlugin = require( 'clean-webpack-plugin' );
 const CopyWebpackPlugin = require( 'copy-webpack-plugin' );
