+++ config/webpack.prod.js
@@ -1,5 +1,3 @@
-/* global module, require */
-
 const merge = require( 'webpack-merge' );
 const common = require( './webpack.common.js' );
 const TerserPlugin = require( 'terser-webpack-plugin' );
