+++ includes/functions/internationalization.php
@@ -16,28 +16,6 @@ function setup() {
 	$n = function( $function ) {
 		return __NAMESPACE__ . "\\$function";
 	};
-
-	add_filter( 'pll_copy_post_metas', $n( 'copy_post_metas' ) );
-}
-
-/**
- * Determine which post meta should be synchronized between translations.
- *
- * @param array $metas The array of metadata keys to synchronize.
- */
-function copy_post_metas( $metas ) {
-	$unset = [
-		// Link will differ for the translated version.
-		'lc_resource_permanent_link',
-		// Translator will be specific to the translated version.
-		'lc_resource_translator',
-	];
-	foreach ( $metas as $key => $value ) {
-		if ( in_array( $value, $unset, true ) ) {
-			unset( $metas[ $key ] );
-		}
-	}
-	return $metas;
 }
 
 /**
